Modularizing server code.
