+++ server/server.js
@@ -13,7 +13,7 @@ var request = require('request');
 // var moment = require('moment');
 var utils = require('./utils.js');
 var documentUtils = require('./documents/documentUtils');
-var userAuthUtil = require('./userProfile/UserSignIn');
+var userAuthUtil = require('./userProfile/userOAuthUtils');
 var socketUtils = require('./socketUtils');
 var userUtils = require('./userProfile/userUtils');
 var path = require('path');
