+++ server/utils.js
@@ -1,13 +1,10 @@
-var dbUser = require('./userProfile/UserModel');
-var dbDocument = require('./documents/DocumentModel');
-var User = dbUser.user;
-var userDocument = dbDocument.userDocument;
 
 module.exports = {
   forceHTTPS: function(req, res, next) {
     var protocol = req.get('x-forwarded-proto');
     protocol == 'https' ? next() : res.redirect('https://' + req.hostname + req.url);
   }, 
+
   allowCrossOrigin: function(req, res, next) {
     res.header("Access-Control-Allow-Origin", "*");
     res.header("Access-Control-Allow-Methods", "PUT, GET, POST, DELETE, OPTIONS");
@@ -25,8 +22,8 @@ module.exports = {
         content = data;
         //To get around the synchronous behavior we wrap the next step into the function sendFileDataToClient. Which will just emit the content, but this way we are sure that content is done receiving the data from the file.
         sendFileDataToClient(content);
-
       }
     });
   }
+  
 };
